backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-cloud-bigquery==3.18.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.43.0
pandas==2.1.4
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
google-cloud-scheduler==2.13.0
